requirements.txt
Flask>=2.0
flask-mysqldb
PyMySQL
python-dotenv
